Moving schema & naming related functions into new file util/Schemas.js
