+++ src/encoded/static/components/util/index.js
@@ -53,3 +53,4 @@ module.exports.DateUtility  = require('./date-utility');
 
 module.exports.expFxn       = require('./experiments-transforms');
 module.exports.Filters      = require('./experiments-filters');
+module.exports.Schemas      = require('./Schemas');
